backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.1.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
yarl==1.22.0
pymongo==4.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gunicorn==22.0.0
